Abstracted Bloom filter into it's own class (and shuffled the test around to handle it).
Still several things out of whack here:

1. Bloom filter is currently only 32 bits wide.  For any non-trivial number of input
values this easily fills up with ones, which in turn causes every "contains" test
to return true.  Filter obviously needs to be much larger, but murmur hash only
supports 32-bit return values... so we need to introduce a mapping between these
values and a larger bit field.
2. We're using a feedback mechanism to generate some number of hash functions;
output from f1 is used as seed value for f2.  Trouble is that the generation
of hash functions is thus tied to the input value, which in turn means that all
inputs are not going through the same set of hash functions.  This SHOULD be a
problem, although there may be some kind of compensation going on that I haven't
considered yet.